+++ programs/ticket-sale/src/account_data/sale.rs
@@ -16,7 +16,7 @@ pub struct Sale {
   pub event_id: u64,
 
   /// A unique index that will differentiate multiple sales of one single event
-  pub ticket_type_index: usize,
+  pub ticket_type_index: u8,
 
   /// The ticket type that decides the sale mechanism
   pub ticket_type: TicketType,
